Backend. Using of H2-db: - add drivers of H2 into Tomcat and correct some configs; - create db-scheme and fill in it with demo-data on server's start by the easiest way; - create dao for albums with one method to test connect to db; - some entities' refactoring for simplifying.

+++ photo-core/src/main/java/com/entity_core/UserInfo.java
@@ -1,7 +1,5 @@
 package com.entity_core;
 
-import java.math.BigInteger;
-
 /**
  * Entity for user's data
  * Created by Aesonne on 02.05.2017
@@ -11,23 +9,23 @@ public class UserInfo {
     /**
      * Identifier
      */
-    private BigInteger id;
+    private Long id;
 
     /**
      * User's nickname. By the way used as a login for simple login process
      */
     private String nick;
 
-    public UserInfo(BigInteger id, String nick) {
+    public UserInfo(Long id, String nick) {
         this.id = id;
         this.nick = nick;
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
